Server

+++ Daejeon_People_Server/src/main/java/com/planb/restful/account/after_signup/FindId_VerifyCode.java
@@ -3,19 +3,20 @@ package com.planb.restful.account.after_signup;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
+import org.json.JSONObject;
+
 import com.planb.support.crypto.AES256;
 import com.planb.support.routing.API;
 import com.planb.support.routing.REST;
 import com.planb.support.routing.Route;
-import com.planb.support.utilities.Mail;
 import com.planb.support.utilities.MySQL;
 
 import io.vertx.core.Handler;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.ext.web.RoutingContext;
 
-@API(functionCategory = "아이디 찾기", summary = "인증번호 확인 후 이메일로 아이디 전송")
-@REST(requestBody = "email : String, code : String", successCode = 201, failureCode = 204, etc = "전송한 이메일 인증 코드가 없을 경우 fail")
+@API(functionCategory = "아이디 찾기", summary = "인증번호 확인 후 아이디 응답")
+@REST(requestBody = "email : String, code : String", responseBody = "id: String", successCode = 201, failureCode = 204, etc = "전송한 이메일 인증 코드가 없을 경우 fail")
 @Route(uri = "/find/id/verify", method = HttpMethod.POST)
 public class FindId_VerifyCode implements Handler<RoutingContext> {
 	@Override
@@ -23,30 +24,31 @@ public class FindId_VerifyCode implements Handler<RoutingContext> {
 		String email = ctx.request().getFormAttribute("email");
 		String code = ctx.request().getFormAttribute("code");
 		
-		ctx.response().setStatusCode(findIdVerify(email, code)).end();
-		ctx.response().close();
-	}
-	
-	private int findIdVerify(String email, String code) {
 		String encryptedEmail = AES256.encrypt(email);
 		
 		ResultSet rs = MySQL.executeQuery("SELECT * FROM email_verify_codes WHERE email=? AND code=?", encryptedEmail, code);
 		try {
-			if(rs.next()) {
+			if(rs != null && rs.next()) {
 				// 전송한 인증 코드가 있을 경우
 				
 				MySQL.executeUpdate("DELETE FROM email_verify_codes WHERE email=? AND code=?", encryptedEmail, code);
 				rs = MySQL.executeQuery("SELECT * FROM account WHERE email=?", encryptedEmail);
+				assert rs != null;
 				rs.next();
-				String decryptedId = AES256.decrypt(rs.getString("id"));
-				Mail.sendMail(email, "[대전사람] 아이디 찾기 결과입니다.", "ID : " + decryptedId);
-				return 201;
+				
+				String id = rs.getString("id");
+				
+				JSONObject response = new JSONObject();
+				response.put("id", id);
+				
+				ctx.response().setStatusCode(201);
+				ctx.response().end(response.toString());
 			} else {
-				return 204;
+				ctx.response().setStatusCode(204).end();
+				ctx.response().close();
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
-			return 204;
 		}
 	}
 }
